utilities package created, added a utility class
